+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Make some additional functions and modules private, including the `vector` module.
 - Update docstrings to improve formatting of sphinx auto-docs.
+- Rename args to `time_average_z_poynting_flux` from `*_fields` to `*_field` to be consistent with other functions in the `fields` module.
 
 ## 0.14.1 (March 4, 2025)
 - Set `Formulation.JONES_DIRECT_FOURIER` as the default formulation for all eigensolve functions in `fmm` module.